jobboard-backend/requirements.txt
Django
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
drf-yasg
python-dotenv
gunicorn
django-cors-headers
django-filter
django-seed